fix(clippy): resolve -D warnings failures in daemon and cli
- query_expand: use `?` instead of manual is_none/return
- search_cache: replace contains_key+insert with get_mut
- service: allow too_many_arguments on hybrid_search_tiered
- cli: drop unnecessary cast, redundant format borrow, allow write_text_brief

+++ crates/statefulmemory-cli/src/cmd_ingest.rs
@@ -61,7 +61,7 @@ pub async fn dispatch(
         })
         .unwrap_or_else(|| DEFAULT_EXTS.iter().map(|s| (*s).to_string()).collect());
 
-    let max_bytes = a.max_bytes.unwrap_or(256 * 1024) as u64;
+    let max_bytes = a.max_bytes.unwrap_or(256 * 1024);
     let session_id = a
         .session
         .clone()
